fix(collection): check if file existed before windows replace

+++ mud/collection.py
@@ -8,6 +8,7 @@ import glob
 import json
 import logging
 import os
+import os.path
 import settings
 
 
@@ -76,12 +77,22 @@ class FileStorage(CollectionStorage):
         with open(temp_path, "w") as fh:
             fh.write(json.dumps(record, indent=4, sort_keys=True))
 
-        windows_temp_path = temp_path + ".WIN"
-        os.rename(path, windows_temp_path)
+        # Check for Windows to perform a two-step rename
+        # instead of overwriting target file
+        path_exists = os.path.exists(path)
+
+        if path_exists:
+            windows_temp_path = temp_path + ".WIN"
+
+            if os.path.exists(windows_temp_path):
+                os.remove(windows_temp_path)
+
+            os.rename(path, windows_temp_path)
 
         os.rename(temp_path, path)
 
-        os.remove(windows_temp_path)
+        if path_exists:
+            os.remove(windows_temp_path)
 
     def post_delete(self, record):
         path = self.get_record_path(record)
